Remove unused abstract base class: SessionRequests is only used by the LocalNewSessionQueue
